+++ CHANGES.md
@@ -1,2 +1,6 @@
+# v0.2.0 (2023-10-07)
+* Pin dependency versions in `producer.opam` and `dune-project`
+* Add support for monads with two type parameters (e.g. `Result`)
+
 # v0.1.0 (2023-10-04)
 * Initial release